+++ Utils/memory.gd
@@ -9,8 +9,8 @@ func _save() -> void:
 		if !node.has_method("save"):
 			print("Node '%s' is missing a save function, skipped" % node.name)
 		else: 
-			var node_data = node.call("save")
-			save_file.store_line(JSON.stringify(node_data))
+			var data = node.call("save");
+			save_file.store_line(JSON.stringify(data))
 	save_file.close();
 
 func _load() -> void:
@@ -23,10 +23,13 @@ func _load() -> void:
 	while save_file.get_position() < save_file.get_length():
 		var json = JSON.new();
 		json.parse(save_file.get_line())
-		var node_data = json.get_data();
-		if(node_data.has("path") && has_node(node_data["path"])):
-			get_node(node_data["path"]).load(node_data);
-		elif(node_data.has("player")):GLOBAL.player_data_to_load = node_data;
+		var data = json.get_data();
+		if(data.has("path") && has_node(data["path"])):
+			var node: Node = get_node(data["path"]);
+			if(node.has_method("load")): node.load(data);
+			else: 
+				print("Node '%s' is missing a load function, skipped" % node.name)
+		elif(data.has("player")):GLOBAL.player_data_to_load = data;
 	GLOBAL.no_saved_data = false;
 	save_file.close();